requirements.txt
streamlit>=1.36
ultralytics==8.3.202
opencv-python-headless>=4.9.0
numpy>=1.26
pillow>=10.3
